+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 826
+/// Strings: 830
 ///
-/// Built on 2026-06-12 at 21:41 UTC
+/// Built on 2026-06-13 at 08:11 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import